MOSC-1418 CNIS

+++ database/2.5.0/07_atualizar_osc.sql
@@ -15,7 +15,7 @@ DECLARE
 	dado_anterior RECORD;
 	dado_posterior RECORD;
 	flag_update BOOLEAN;
-	osc NUMERIC;
+	osc INTEGER;
 
 BEGIN
 	nome_tabela := 'osc.tb_osc';
